Move swtpm TCTI to a public ConnectToSwtpm helper in swtpm.go
The socket-based TCTI and connection logic was defined inside
tpmea_test.go, so consumers of the library who also want to run against
swtpm had to reimplement it. Moved the whole thing into swtpm.go as a
public ConnectToSwtpm function and simplified TestMain to just call it.
