+++ tpmea_test.go
@@ -7,16 +7,11 @@ import (
 	"crypto/elliptic"
 	crand "crypto/rand"
 	"crypto/rsa"
-	"encoding/binary"
 	"errors"
-	"fmt"
-	"io"
 	"math/rand"
-	"net"
 	"os"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/canonical/go-tpm2"
 )
@@ -29,55 +24,11 @@ const (
 
 var PCR_INDEXES = []int{0, 1, 2, 3, 4, 5}
 
-type swtpmSocketTCTI struct {
-	conn net.Conn
-	rsp  *bytes.Reader
-}
-
 func TestMain(m *testing.M) {
 	if socketPath := os.Getenv("SWTPM_SERVER_PATH"); socketPath != "" {
-		getTpmHandle = func() (*tpm2.TPMContext, error) {
-			conn, err := net.Dial("unix", socketPath)
-			if err != nil {
-				return nil, fmt.Errorf("cannot connect to swtpm socket %s: %w", socketPath, err)
-			}
-			return tpm2.NewTPMContext(&swtpmSocketTCTI{conn: conn}), nil
-		}
+		ConnectToSwtpm(socketPath)
 	}
 	os.Exit(m.Run())
-}
-
-func (t *swtpmSocketTCTI) Read(data []byte) (int, error) {
-	if t.rsp == nil {
-		hdr := make([]byte, 6)
-		if _, err := io.ReadFull(t.conn, hdr); err != nil {
-			return 0, err
-		}
-		responseSize := binary.BigEndian.Uint32(hdr[2:6])
-		buf := make([]byte, responseSize)
-		copy(buf, hdr)
-		if _, err := io.ReadFull(t.conn, buf[6:]); err != nil {
-			return 0, err
-		}
-		t.rsp = bytes.NewReader(buf)
-	}
-	n, err := t.rsp.Read(data)
-	if err == io.EOF {
-		t.rsp = nil
-	}
-	return n, err
-}
-
-func (t *swtpmSocketTCTI) Write(data []byte) (int, error) { return t.conn.Write(data) }
-func (t *swtpmSocketTCTI) Close() error                   { return t.conn.Close() }
-func (t *swtpmSocketTCTI) MakeSticky(_ tpm2.Handle, _ bool) error {
-	return errors.New("not implemented")
-}
-func (t *swtpmSocketTCTI) SetTimeout(timeout time.Duration) error {
-	if timeout == tpm2.InfiniteTimeout {
-		return t.conn.SetDeadline(time.Time{})
-	}
-	return t.conn.SetDeadline(time.Now().Add(timeout))
 }
 
 // genTpmKeyPairRSA generates a 2048 bit RSA key,
